+++ src/config.ts
@@ -1,0 +1,11 @@
+const config = {
+    REPL_FILE_EXT: 'repl'
+};
+
+type ReplSessionType = 'clj' | 'cljs';
+
+export {
+    ReplSessionType
+}
+
+export default config;